Introduce notion of rawConfig, which we use for the options controls.
Before, we were using config, but it was normalized, which could have made all kinds of differences from the underlying data (e.g. adding an individuals [] array if none existed). That caused the weird behavior where individuals couldn't be removed... but also was a more serious problem if there was more kooky normalization going on, because it could create nonsensical edits to the underlying data. Part of #25.