Change the signature of simulator.generator from (previousFrames, simOptions, rnd, runIndex) to (frameIndex, previousFrame, simOptions, rnd, runIndex).
The previous signature required all frames to be kept in memory and passed into the generator. This was fine for small simulators, but would get oppressive for simulators that might have thousands of frames.

This new signature allows flexibility; for example for simulators with very large numbers of frames, we could keep just the last frame... or keep every 50th frame, and if you need a frame like 25, could just recalculate frame 0..25 .

This also makes the common case of "I just want the previousFrame" easier. One downside is that generators that DO need access to earlier frames will need new machinery--likely some kind of indirect getter that can fetch a previous frame, but only one smaller than frameIndex.

Part of #23.