Change the format of config files.
Before it was just a top-level array of configs. However, that is a hard to extend format--what if we need to pack additional top-level information, like version number?

So move it to an object with a small number of keys and validation. This is important for #32, where there might be old versions stored that have to be detected. And #32 could use a different way of storing base configs (MAYBE).