The major bit of work to address #44. packModificationsForUrl, and unpackModifiationsFromUrl now will modify simulator as they're going.
This allows them to correctly pack and unpack short names even when the simulator switches mid diff. They can do this because they can assume all of the necessary simulators are now loaded.

The bug that motivated #44 is now fixed... although the machinery throughout is now much more complex and there are still some lingering oddities captured in that bug.