Make it so PositionedGraphRenderer will work for frames with no graphs.
In that case, it will fall back to agent.x and agent.y. If those aren't set it will fall back to a seeded random x,y based on the id of the agent.

It basically recreates the override points from PositionedGraph, but on PositionedGraphRenderer for an agent.

Part of #66.