A new attempt that does successfully extract the function information.
It does so by mangling the input files to remove anything below `export default` line in the simualtor file,
which by convention cuts off all of the imports that include renderers.

This is a kind of hacky approach, but... it does kind of work? :shrug:

src/simulator.ts : BaseRenderer had to be modified because if it returned BaseRenderer() then it triggered
an error in the mangled files.

Part of #73.