Create aa `npm run generate:stubs` to just generate type stubs. And run it in the build command.
This helps ensure that in a checkout with no GENERATED stubs, the build can at least work, which is
necessary before running `npm run start` or `npm run build`, which is necessary to run the full `npm run
generate flow.

Also add a `npm run generate:pre-build` and include it in the build step.

Part of #73.
 # Please enter the commit message for your changes. Lines starting