Make it so CoordiantesMapFrameData stores just an array of the IDs, and not the fullData.
It also stores a hash to verify that the fullitems it was passed are the ones it was working with.

This will make it store considerarbly less duplicative data in the common case, and help make sure it
notices when it's working with data that differs from when it was last saved.

Part of #87.