Move to a flow that generates the emoji constants automatically via `npm run generate:internal:emojis`
Currently it's a semantic no-op but soon it will be useful for generating gendered emoji.

You can't use `as const` type without it being an object literal, which means that automatically generating
e.g. gendered and skin toned versions of people emoji would break the explicit enumerations of KnownEmoji*.

Part of #68.